a better commit description :D still need to get a list of the dev libs up, added skybox to render r_skybox 0/1 enables disables textures are in id1/gfx/env along with original authors readme.
revamped mutliplayer menu,
no longer goes to the direct/modem/ipx/tcp-ip menu
it now goes directly to join/ new/setup.

setup in multiplayer now has a device choice menu item,
in the device choice menu it lists all of the active&&has ip address interfaces.
of which you may choose which one to use.
this is limited to the first five that it finds.

the  game will bind the udp socket to the last active&&has ip address interface
it finds.
this should be fine for normal users.

the toolsrc folder is new,
it has the original qutils atm, all get built with the engine via cmake,
just remove the addsubdirectory entry from the top level cmake if need be.

oh yes,

extlib now contains android-iffaddrs,
this is used in the afor mentioned network changes,
and to make sure there is android compatability when i
start working on building for android.

i dont think ive forgoten any thing,
ill try to be more descriptive in my future commits.